pyproject.toml
[tool.poetry]
name = "BancoDeDadosTrabFInal"
version = "0.1.0"
description = ""
authors = ["Arthur <munhozarthur@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg = "^3.0.11"
jsonschema = "^4.4.0"
develop = "^0.2.1"
black = "^22.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
